+++ test/dummy/app/models/message.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
 class Message < ApplicationRecord
+  validates :title, presence: true
+
   has_draft :title, :content
 end