+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.2.0
+-----
+
+- A `publish` method for assignment the draft values to actual attributes.
+- Fallback to actual attribute value when draft is nil.
+
 0.1.0
 -----